+++ fabric/src/main/java/io/github/sakurawald/fuji/core/config/model/ConfigModel.java
@@ -60,6 +60,19 @@ public class ConfigModel {
             public String logger_level = "WARN";
         }
 
+
+        public DocumentSection document = new DocumentSection();
+        public static class DocumentSection {
+            @Document(id = 1752484948715L, value = """
+                When `enable` this option, we will always use the `built-in doc strings`, which is written in `English`.
+                This option is used to ensure you always sees the `latest version` of `doc strings` from your `current using version`.
+
+                When `disable` this option, we will ues the `doc strings` from the `language files`.
+                """)
+            public boolean alwaysUseBuiltInDocStrings = true;
+        }
+
+
         public static class Backup {
 
             @Document(id = 1751823774480L, value = """